requirements.txt
tensorflow
scikit-learn
numpy
matplotlib
seaborn
fastapi
streamlit
uvicorn
pillow
python-multipart